Update README.md
The objective of this project is to develop a text extraction and cleaning tool that utilizes the capabilities of Amazon Textract and RapidMiner. The code extracts text from documents stored in an AWS S3 bucket, performs text cleaning operations using RapidMiner, and saves the extracted and cleaned text to Excel files.

This project enables efficient text extraction and cleaning from documents stored in an AWS S3 bucket. By leveraging Amazon Textract for extraction and RapidMiner for cleaning, it provides a comprehensive solution for processing textual data. The extracted and cleaned text can be further utilized for various applications, such as data analysis, natural language processing, and information retrieval.